refactor: signal_adapter infrastructure/adapters/ → mcp/ 이동
proxy-cleanup Step C: signal_adapter.py를 core/mcp/signal_adapter.py로 이동.
tests/test_signal_port.py import 경로 갱신. 행동 변경 없음.
