+++ core/mcp/signal_adapter.py
