+++ tests/test_signal_port.py
@@ -5,7 +5,7 @@ from __future__ import annotations
 from typing import Any
 
 from core.domains.game_ip.nodes.signals import set_signal_adapter, signals_node
-from core.infrastructure.adapters.signal_adapter import (
+from core.mcp.signal_adapter import (
     FixtureSignalAdapter,
     LiveSignalAdapter,
     create_signal_adapter,
